+++ packages/uui-table/lib/uui-table.test.ts
@@ -70,7 +70,7 @@ describe('UuiTable', () => {
     expect(row.selected).to.be.false;
   });
 
-  it('passes the a11y audit', done => {
-    expect(table).shadowDom.to.be.accessible({ done });
+  it('passes the a11y audit', async () => {
+    await expect(table).shadowDom.to.be.accessible();
   });
 });